[~] Some refactoring
